Remove HearthDb projects and update references
The Visual Studio solution file (`hs.sln`) has been updated to remove entries for the `HearthDb`, `HearthDb.CardDefsDownloader`, `HearthDb.CardIdGenerator`, and `HearthDb.EnumsGenerator` projects, along with their configuration settings.

In the `raw.csproj` file, the `<PlatformTarget>` and debug properties have been removed, and a new `<Platforms>` property with the value `AnyCPU` has been added. The project reference to `HearthDb` has been replaced with a folder entry for `Properties`, and a new reference to `HearthDb.dll` in the Release directory has been added. The reference to `HearthMirror` has also been updated to a new location in the `lib` directory.